Address feedback and fix potential undefined behavior in seqreader.cc
